Change site deatails path to allow for params
